gitconfig: add subtree-fetch alias
This alias does a fetch of the specified remote (or 'origin'), and then,
if remote.$remote.subtree is specified, does a filter-branch of the
remote refs with a subdirectory-filter set to that subtree. This is
useful if, say, you need to subtree merge a subdirectory of a remote
branch, rather than the root.

Example setup:
git remote add other_repo git://git.busybox.net/busybox
git config remote.other_repo.subtree mailutils/
git subtree-fetch other_repo

Now you can do a subtree merge of other_repo/master, and you'll only be
pulling over the contents of the mailutils/ directory.

Signed-off-by: Chris Larson <chris_larson@mentor.com>