Add git-clone-via script
Clones from one local repository to another, but without making the
source the remote of the destination. Instead, the destination gets all
the bits from the source's origin, and gets its remote redirected back
to the upstream which the source points at.

In this way, you essentially get a new clone from upstream, but with
hard linking from the existing repository's objects, and you get a
functional repository even if upstream is down, or your network is
unavailable.

Signed-off-by: Chris Larson <chris_larson@mentor.com>