update-gist-generic: add script
This script accepts content on stdin and a gist url and filename as
commandline arguments. The stdin content is diffed against the current content
in the gist, this diff is shown to the user, then the gist is updated with
that content. If stdout is a tty, the user will be prompted about whether they
want to continue the update, and a -n argument for a dry run is provided. I've
used this pattern in a number of my scripts, so it's now generic.

Signed-off-by: Christopher Larson <kergoth@gmail.com>