git: improve the 'dangling' alias
Rather than showing a list of commit hashes, show a nice git log-like display.
Patches are shown by default, as it fits the use case -- if I'm looking at
dangling commits, I want to know precisely what they are, as I'm checking to
see if something needs to be resurrected. Log format is used rather than
lg-format for the same reason. lgd is now redundant, so is just an alias for
dangling.

Signed-off-by: Christopher Larson <kergoth@gmail.com>