git-prune-merged: add 'corresponding' option
This prunes the branches which have been merged to the branch of the
same name in the corresponding remote, i.e. the 'stable' branch is
pruned if it's fully merged into the 'stable' branch upstream, rather
than comparing all the branches to a single upstream branch. This is
useful for forks.

Signed-off-by: Christopher Larson <kergoth@gmail.com>