zsh: add fallback promptline when base16 color changes don't work
Specifically disable it for vscode and Apple_Terminal for now. It's
possible there's a way to query the terminal to determine if its capable
of modifying colors with escape codes, but I don't know how to do that
at this time, so this will do.

The fallback is generated from the old vim baycomb scheme.

Signed-off-by: Christopher Larson <kergoth@gmail.com>