vim: drop 'jj' imap
The intended purpose of this mapping is to avoid reaching so far to the
escape key, but on all the machines I use, I map capslock to escape on
press and control on hold, so I never reach up to escape anyway. As
a result, I never use this mapping.

Signed-off-by: Christopher Larson <kergoth@gmail.com>