Revert "os-wsl: sanity check WSL_AUTH_SOCK"
This reverts commit 8fd451b514573b560281e17d9136359c4b0fd0c0.

This causes some pain in my case, as the windows ssh agent will refuse
connections when I'm not on the VPN due to this machine being on the
domain. With this applied, I had to restart my shell to pick up
SSH_AUTH_SOCK after connecting to the VPN, so revert it to avoid this.