Set window title for PowerShell
This behavior aligns with my zsh config, which sets the window title to
the current command when a command is running, and the current path
otherwise. This uses a PSReadLine key handler to set the title based on
the command, and a starship precommand hook to set the title based on
the current path.

Close #15

Signed-off-by: Christopher Larson <kergoth@gmail.com>