Rework the deck backup scripts
These now have a single entry point which calls out to individual backup job scripts. This allows for more flexibility in the future, such as adding new backup jobs, or running individual jobs on demand. This also shifts the logic for running a job based on frequency since last execution into the backup script, rather than the separate check-backup script, with an additional argument to force a backup run. The systemd backup service forces it to run for all jobs.