Skip 1password cli installation if the age key is present
Also align the logic with the age key bootstrap scripts for consistency.

I considered merging the two scripts, but the complexity due to the evaluation order with use of the template 1password functions was too high. I did consider calling the `op` command directly, but then we would need to ensure we obey the chezmoi arguments relating to interactivity and force mode, so this is simpler.

Close #34